requirements.txt
# --------- pytorch --------- #
torch==1.12.0
torchvision==0.13.0
torchaudio==0.12.0
torchtext==0.13.0
pytorch-lightning==1.7.1
torchmetrics==0.10.0

# --------- hydra --------- #
hydra-core==1.2.0
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
# wandb
# neptune-client
# mlflow
# comet-ml

# --------- others --------- #
pyrootutils     # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
nnAudio==0.3.1
# sh            # for running bash commands in some tests (linux/macos only)
